+++ src/test/java/es/princip/getp/domain/auth/service/EmailVerificationServiceTest.java
@@ -15,7 +15,7 @@ import org.mockito.junit.jupiter.MockitoExtension;
 
 import java.util.Optional;
 
-import static es.princip.getp.fixture.EmailVerificationFixture.createEmailVerification;
+import static es.princip.getp.domain.auth.fixture.EmailVerificationFixture.createEmailVerification;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.mockito.ArgumentMatchers.any;
