+++ src/test/java/es/princip/getp/domain/member/service/MemberServiceTest.java
@@ -6,8 +6,8 @@ import es.princip.getp.domain.member.dto.request.CreateMemberRequest;
 import es.princip.getp.domain.member.dto.request.UpdateMemberRequest;
 import es.princip.getp.domain.member.exception.MemberErrorCode;
 import es.princip.getp.domain.member.repository.MemberRepository;
+import es.princip.getp.domain.serviceTerm.fixture.ServiceTermFixture;
 import es.princip.getp.domain.serviceTerm.service.ServiceTermService;
-import es.princip.getp.domain.serviceTerm.support.ServiceTermFixture;
 import es.princip.getp.domain.storage.service.ImageStorageService;
 import es.princip.getp.global.exception.BusinessLogicException;
 import es.princip.getp.global.util.PathUtil;
